==> 00_clean_keywords.ipynb <==
import re

with open('data/keywords.txt') as f:
    lines = f.readlines()

kw = [elem.strip() for elem in lines]

pattern = r'"(.*?)"'

kw = [re.match(pattern, elem).group(1) if re.match(pattern, elem) else elem for elem in kw]

len(kw)

kw_set = []
for w in kw:
    if w not in kw_set:
        kw_set.append(w)

len(kw_set)

kw_set

to_remove = ['(ayuda',
             'donados',
             'productos) AND (vendida',
             'desviados',
             'revende',
             'comercializada)']

kw_set = [w for w in kw_set if w not in to_remove]

len(kw_set)

kw_set = [re.sub('\(|\)', '', w) for w in kw_set]

kw_set

import pandas as pd
df = pd.DataFrame(kw_set, columns=['Keywords'])

df.to_csv('data/keywords_clean.csv')
