==> 02_clean_tweets.ipynb <==
import json
import os
import pandas as pd

dump_path = 'data/tweets/'

dirs = [dump_path + f + '/' for f in os.listdir(dump_path) if f[0].isnumeric()]

dirs = [d for d in dirs if os.listdir(d)]

dirs

all_tweets = []

for d in dirs:
    with open(d + 'clean_tweets.json', 'rb') as fin:
        res = json.load(fin)
        all_tweets += res

len(all_tweets)

id_tweet_dict = {}

for tweet in all_tweets:
    if tweet['id'] in id_tweet_dict.keys():
        id_tweet_dict[tweet['id']].append(tweet)
    else:
        id_tweet_dict[tweet['id']] = [tweet]

len(id_tweet_dict)

res = [v[0] for v in id_tweet_dict.values()]

with open('data/tweets/_all_cleaned/tweets.json', 'w') as fout:
    json.dump(res, fout)

pd.DataFrame(res).to_parquet('data/tweets/_all_cleaned/tweets.parquet')

import json
import os
import pandas as pd

dump_path = 'data/tweets/'

dirs = [dump_path + f + '/' for f in os.listdir(dump_path) if f[0].isnumeric()]

dirs = [d for d in dirs if os.listdir(d)]

dirs

all_users = []

for d in dirs:
    with open(d + 'clean_users.json', 'rb') as fin:
        res = json.load(fin)
        all_users += res

len(all_users)

all_users[0]

id_user_dict = {}

for user in all_users:
    if user['id'] in id_user_dict.keys():
        id_user_dict[user['id']].append(user)
    else:
        id_user_dict[user['id']] = [user]

len(id_user_dict)

res = [v[0] for v in id_user_dict.values()]

with open('data/tweets/_all_cleaned/users.json', 'w') as fout:
    json.dump(res, fout)

pd.DataFrame(res).to_parquet('data/tweets/_all_cleaned/users.parquet')
