==> 04a_news_get_tweets.ipynb <==
# Get News Tweets

import pandas as pd
import scraper
import json

from datetime import datetime

news_accounts = pd.read_csv('data/news/twitter_users_news.csv', header=None, names=['handle', 'id', 'website'])
news_accounts

news_accounts.dtypes

start_date = datetime.strptime('2022-03-16', '%Y-%m-%d')

end_date = datetime.strptime('2022-04-01', '%Y-%m-%d')

dump_path = f'data/news/news_tweets/{start_date.strftime("%Y-%m-%d")}__{end_date.strftime("%Y-%m-%d")}/'

dump_path

client = scraper.create_client(scraper.init())

tweets, users = scraper.get_users_tweets(client=client,
                                         user_ids=news_accounts['id'].tolist())

with open(dump_path + 'tweets.json', 'w') as fout:
    json.dump(tweets, fout)

with open(dump_path + 'users.json', 'w') as fout:
    json.dump(users, fout)
